// src/types/addPlaceNames.ts
export type addPlaceNameElement = {
  id: string;
  name: string;
};

export type addPlaceName = {
  basic: addPlaceNameElement;
  location: addPlaceNameElement;
  description: addPlaceNameElement;
  mainPhotos: addPlaceNameElement;
  area: addPlaceNameElement;
  details: addPlaceNameElement;
  weddingAlbum: addPlaceNameElement;
};